requirements.txt
Flask==2.3.3
numpy==1.24.3
pandas==2.0.3
sentence-transformers==2.2.2
scikit-learn==1.3.0
PyPDF2==3.0.1
Werkzeug==2.3.7
Pillow==10.0.0
gunicorn==21.2.0